Move all view setups from MainActivity to MainViewMvcImpl
